Remove selector tests
